[Docs] Move each demo into it folder
