+++ docs/src/Main.native.js
@@ -6,11 +6,15 @@ import React, {
   ScrollView,
 } from 'react-native';
 
-import SwipeableViews from '../../src/index.native.animated';
 import Head from './Head.native';
 import Body from './Body.native';
-import Demo from './Demo.native';
-import Demo2 from './Demo2.native';
+import Demo from './Demo/Demo.native';
+import Demo1 from './Demo/Demo1.native';
+import Demo2 from './Demo/Demo2.native';
+import Demo3 from './Demo/Demo3.native';
+import Demo4 from './Demo/Demo4.native';
+import Demo5 from './Demo/Demo5.native';
+import Demo6 from './Demo/Demo6.native';
 
 const styles = StyleSheet.create({
   root: {
@@ -69,23 +73,7 @@ const Main = () => (
         name="Demo 1"
         description="Simple case without header."
       >
-        <SwipeableViews containerStyle={styles.slideContainer}>
-          <View style={[styles.slide, styles.slide1]}>
-            <Text style={styles.text}>
-              slide n°1
-            </Text>
-          </View>
-          <View style={[styles.slide, styles.slide2]}>
-            <Text style={styles.text}>
-              slide n°2
-            </Text>
-          </View>
-          <View style={[styles.slide, styles.slide3]}>
-            <Text style={styles.text}>
-              slide n°3
-            </Text>
-          </View>
-        </SwipeableViews>
+        <Demo1 styles={styles} />
       </Demo>
       <Demo
         name="Demo 2"
@@ -97,89 +85,25 @@ const Main = () => (
         name="Demo 3"
         description="Set a constant height and let the swipe and scroll behavior work in harmony."
       >
-        <SwipeableViews containerStyle={styles.slideContainer}>
-          <ScrollView style={[styles.slide, styles.slide1]}>
-            {list}
-          </ScrollView>
-          <View style={[styles.slide, styles.slide2]}>
-            <Text style={styles.text}>
-              slide n°2
-            </Text>
-          </View>
-          <View style={[styles.slide, styles.slide3]}>
-            <Text style={styles.text}>
-              slide n°3
-            </Text>
-          </View>
-        </SwipeableViews>
+        <Demo3 styles={styles} list={list} />
       </Demo>
       <Demo
         name="Demo 4"
         description="Or let the container respond dynamically to its children."
       >
-        <SwipeableViews>
-          <View style={[styles.slide, styles.slide1]}>
-            {list.slice(0, 10)}
-          </View>
-          <View style={[styles.slide, styles.slide2]}>
-            {list.slice(0, 7)}
-          </View>
-          <View style={[styles.slide, styles.slide3]}>
-            {list.slice(0, 3)}
-          </View>
-        </SwipeableViews>
+        <Demo4 styles={styles} list={list} />
       </Demo>
       <Demo
         name="Demo 5"
         description="Add a resistance bounds effet on the edges."
       >
-        <SwipeableViews containerStyle={styles.slideContainer} resistance={true}>
-          <View style={[styles.slide, styles.slide1]}>
-            <Text style={styles.text}>
-              slide n°1
-            </Text>
-          </View>
-          <View style={[styles.slide, styles.slide2]}>
-            <Text style={styles.text}>
-              slide n°2
-            </Text>
-          </View>
-          <View style={[styles.slide, styles.slide3]}>
-            <Text style={styles.text}>
-              slide n°3
-            </Text>
-          </View>
-        </SwipeableViews>
+        <Demo5 styles={styles} />
       </Demo>
       <Demo
         name="Demo 6"
         description="You can also nest this component."
       >
-        <SwipeableViews containerStyle={styles.slideContainer}>
-          <View style={[styles.slide, styles.slide1]}>
-            <Text style={styles.text}>
-              slide n°1
-            </Text>
-            <View style={styles.divider} />
-            <SwipeableViews resistance={true}>
-              <View style={[styles.slide, styles.slide2]}>
-                <Text style={styles.text}>
-                  slide n°1
-                </Text>
-              </View>
-              <View style={[styles.slide, styles.slide3]}>
-                <Text style={styles.text}>
-                  slide n°2
-                </Text>
-              </View>
-            </SwipeableViews>
-          </View>
-          <View style={[styles.slide, styles.slide2]}>
-            <Text style={styles.text}>
-              slide n°2
-            </Text>
-          </View>
-        </SwipeableViews>
+        <Demo6 styles={styles} />
       </Demo>
     </Body>
   </ScrollView>
